% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alltransf.R
\name{alltransf}
\alias{alltransf}
\title{Computes all the chosen transformations to the input data frame in the chosen
parameters.}
\usage{
alltransf(data, subset, select = colnames(data), transf = c("rsqrt", "log",
  "sqrt"))
}
\arguments{
\item{data}{A data frame}

\item{subset}{a specification of the rows to be used: defaults to all rows.
This can be any valid indexing vector (see \link{[.data.frame}) for the
rows of data or if that is not supplied, a data frame made up of the
variables used in \code{formula}.}

\item{select}{expression, indicating columns to select from a data frame.
Defaults for all the variables in data. See \link{subset}.}

\item{transf}{The chosen transformations to be applied to the data}
}
\value{
A list with all data transformed according to the arguments
  passed.help
}
\description{
Computes all the chosen transformations to the input data frame in the chosen
parameters.
}
\examples{

vars <- colnames(centro_2015@data)
alltransf(centro_2015@data, select = vars)
alltransf(centro_2015@data, 1:10, c("valor", "area_total"))

vars <- colnames(itacorubi_2015@data)
alltransf(itacorubi_2015@data, select = vars)
}
