% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros_nbr14653.R
\name{amplitude}
\alias{amplitude}
\title{Based on a dataframe containing the estimated values, returns amplitude
according to NBR14.653-2}
\usage{
amplitude(Y)
}
\arguments{
\item{Y}{dataframe with \emph{Y_hat} values with confidence/prediction
\code{\link{bestfit}}}
}
\value{
a vector containing the amplitudes of the \emph{Y_hat} in \%
}
\description{
Amplitude according to NBR14.653-2.
}
\examples{
fit <- lm(log(Valor_Total) ~ . - lat - lon, data = centro_2015)
aval <- grepl("^aval", rownames(centro_2015))
Y_hat <- predict(object = fit, interval = "confidence", newdata = centro_2015[aval,])
Y_hat <- inverse(Y_hat, "log")
amplitude(Y_hat)
}
