% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{AREA_CONSTRUIDA}
\alias{AREA_CONSTRUIDA}
\title{Building Area}
\usage{
AREA_CONSTRUIDA(area_terreno, IA, np)
}
\arguments{
\item{area_terreno}{land area}

\item{IA}{rate of land use}

\item{np}{floors}
}
\value{
the maximum building area
}
\description{
Building Area
}
\examples{
ac <- AREA_CONSTRUIDA(630, 2.5, 5)
}
