% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atibaia}
\alias{atibaia}
\title{Different relations between sales prices and listing prices}
\format{
A tibble with 65 rows and 17 variables:
\itemize{
  \item Id: id
  \item Endereco: house address
  \item Descricao: Advertising text
  \item Bairro: neighborhood
  \item AreaConstruida: building area
  \item DataTransacao: sales date
  \item Data: sales date, with convenient format for modelling
  \item Localizacao: indicative of house to be inside gated community or not
  \item IndiceFiscal: land price per square meter
  \item PadraoConstrutivo: construction
  \item OrigemComprador: buyer origin
  \item FinalidadeCompra: purpose of purchase
  \item ValorOfertado: List prices
  \item ValorVendido: Sale prices
  \item FatorOferta: relation between list and sales prices
  \item PUvenda: sale price per square meter
  \item PUoferta: list price per square meter
}
}
\usage{
atibaia
}
\description{
A tibble containing a sample of 65 houses in Atibaia/SP with listing and
sales prices.
}
\examples{
data(atibaia)
fitOferta <- lm(log(PUoferta) ~ log(IndiceFiscal) + log(AreaConstruida) +
                 PadraoConstrutivo, data = atibaia)
fitVenda <- lm(log(PUvenda) ~ log(IndiceFiscal) + log(AreaConstruida) +
                PadraoConstrutivo, data = atibaia)
plotModel(fitOferta, interval = "both", ca = TRUE,
          residuals = TRUE, colour = PadraoConstrutivo)
plotModel(fitVenda, interval = "both", ca = TRUE,
          residuals = TRUE, colour = PadraoConstrutivo)
}
\keyword{datasets}
