% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros_nbr14653.R
\name{aval_intervalar}
\alias{aval_intervalar}
\title{Based on a dataframe containing the estimated values, returns interval evaluation
according to NBR14.653-2}
\usage{
aval_intervalar(Y)
}
\arguments{
\item{Y}{dataframe with \emph{Y_hat} values with or without confidence/prediction 
\code{\link{bestfit}}}
}
\value{
a vector containing interval evaluation according to NBR14.653-2.
}
\description{
Interval evaluation based on NBR14.653-2.
}
\examples{
fit <- lm(log(Valor_Total) ~ . - lat - lon, data = centro_2015)
aval <- grepl("^aval", rownames(centro_2015))
Y_hat <- predict(object = fit, interval = "confidence",
                newdata = centro_2015[aval,])
Y_hat <- inverse(Y_hat, "log")
aval_intervalar(Y_hat)

}
