% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxplot.R, R/plotdf.R
\name{bboxplot}
\alias{bboxplot}
\alias{bboxplot.default}
\alias{bboxplot.formula}
\alias{bboxplot}
\alias{bboxplot.default}
\alias{bboxplot.formula}
\title{Beautiful boxplots with ggplot2.}
\usage{
bboxplot(y, g, data)

\method{bboxplot}{default}(y, g, data)

\method{bboxplot}{formula}(formula, data)

bboxplot(y, g, data)

\method{bboxplot}{default}(y, g, data)

\method{bboxplot}{formula}(formula, data)
}
\arguments{
\item{y}{a numeric variable for which the boxplot is to be constructed.}

\item{g}{a grouping variable, usually a factor, for constructing parallel
boxplots.}

\item{data}{a dataframe containing the variables.}

\item{formula}{A model formula of the form ~ y to produce a boxplot for
the variable y, or of the form y ~ g to produce parallel boxplots for y
within levels of the grouping variable(s) g, etc., usually factors.}

\item{\dots}{further arguments passed to \code{\link[ggplot2]{geom_boxplot}}.}

\item{y}{a numeric variable for which the boxplot is to be constructed.}

\item{g}{a grouping variable, usually a factor, for constructing parallel
boxplots.}

\item{data}{a dataframe containing the variables.}

\item{formula}{A ‘model’ formula, of the form ~ y to produce a boxplot for
the variable y, or of the form y ~ g to produce parallel boxplots for y
within levels of the grouping variable(s) g, etc., usually factors.}

\item{data}{A data frame containing the variables in the formula.}
}
\value{
A boxplot of the variable \code{y} or the boxplot of the \code{y}
  against \code{g}.

A boxplot of the variable \code{y} or the boxplot of the \code{y}
  against \code{g}.
}
\description{
Create boxplots with \link{ggplot2}.

Beautiful boxplots with ggplot2
}
\examples{
data <- as.data.frame(centro_2015@data)
bboxplot("valor", data = data)
bboxplot("valor", "padrao", data)
bboxplot(valor~padrao, data)
bboxplot(~valor, data)
data <- as.data.frame(centro_2015@data)
bboxplot(y = "valor", data = data)
bboxplot(y = "valor", g = "padrao", data = data)

bboxplot(valor ~ padrao, data)
bboxplot(~ valor, data)
}
