% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestfit.R
\name{bestfitEst}
\alias{bestfitEst}
\title{Fitter function for bestfit}
\usage{
bestfitEst(X, y, t, p, response)
}
\arguments{
\item{X}{a design matrix for a regression model}

\item{y}{the vector of the response variable}

\item{t}{The transformed data}

\item{p}{The combinations of transformed variables to be tested}

\item{response}{The name of the response variable}
}
\value{
a vector with adjusted R2 to each fit
}
\description{
This is the basic computing engine called by \code{\link{bestfit}}. This
function should not be used directly by the user.
}
