% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{brformat}
\alias{brformat}
\alias{brf}
\title{Wrapper around format}
\usage{
brformat(
  x,
  decimal.mark = ",",
  big.mark = ".",
  digits = 2,
  nsmall = 2,
  scientific = FALSE,
  ...
)

brf(
  x,
  decimal.mark = ",",
  big.mark = ".",
  digits = 2,
  nsmall = 2,
  scientific = FALSE,
  ...
)
}
\arguments{
\item{x}{a number to be formatted by \code{format}}

\item{digits}{a positive integer indicating how many significant digits
    are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link[base]{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For more, notably the interpretation for complex numbers see
    \code{\link[base]{signif}}.)}

\item{nsmall}{the minimum number of digits to the right of the decimal
    point in formatting real/complex numbers in non-scientific formats.
    Allowed values are \code{0 <= nsmall <= 20}.}

\item{scientific}{Either a logical specifying whether
    elements of a real or complex vector should be encoded in scientific
    format, or an integer penalty (see \code{\link[base]{options}("scipen")}).
    Missing values correspond to the current default penalty.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function is only a wrapper around \link{format} function that uses standard
brazillian formats by default
}
