% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros_nbr14653.R
\name{campo_arbitrio}
\alias{campo_arbitrio}
\title{Based on a dataframe containing the estimated values, returns interval of 
arbitration parameters according to NBR14.653-2}
\usage{
campo_arbitrio(Y)
}
\arguments{
\item{Y}{dataframe with \emph{Y_hat} values with or without
confidence/prediction \code{\link{bestfit}}}
}
\value{
a vector containing the interval of arbitration according to
 NBR14.653-2.
}
\description{
Interval of arbitration according to NBR14.653-2.
}
\examples{
fit <- lm(log(Valor_Total) ~ . - lat - lon, data = centro_2015)
aval <- grepl("^aval", rownames(centro_2015))
Y_hat <- predict(object = fit, interval = "confidence",
                newdata = centro_2015[aval,])
Y_hat <- inverse(Y_hat, "log")
campo_arbitrio(Y_hat)

}
