% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canasvieiras_97}
\alias{canasvieiras_97}
\title{Prices of 15 plots in Canasvieiras (1997)}
\format{a tibble with 15 rows and 5 columns
\itemize{
  \item endereco address.
  \item valor_unitario unitary value per squared meter.
  \item area area in squared meters.
  \item dist_b_mar distance to the sea.
  \item zoneamento zone.
}}
\source{
\strong{HOCHHEIM, Norberto}. \emph{Engenharia de Avaliacoes II}. 
Florianopolis: IBAPE/SC, 2005, p.3.4
}
\usage{
canasvieiras_97
}
\description{
A tibble containing a sample of 15 unitary values of plots sales prices per 
squared meters.
}
\keyword{datasets}
