% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{centre}
\alias{centre}
\alias{centre.numeric}
\alias{centre.factor}
\title{Return the median value or the modal value of a vector, depending on whether
the vector is numeric or a factor.}
\usage{
centre(x, ...)

\method{centre}{numeric}(x, FUN = median, na.rm = TRUE, ...)

\method{centre}{factor}(x, FUN, na.rm = TRUE, ...)
}
\arguments{
\item{x}{an object for which a method has been defined, or a
    numeric vector containing the values whose median is to be computed.}

\item{...}{potentially further arguments for methods; not used in
    the default method.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the median value for objects of the class integer or double. The
  modal value for objects of class factor.
}
\description{
Return the median value or the modal value of a vector, depending on whether
the vector is numeric or a factor.
}
\examples{
x <- c(-3, -2, 0, NA, 1, 1, 3)
centre(x) # .5 (default is median(x))
centre(x, FUN = mean) # 0
centre(x, FUN = raster::modal) # 1
y <- c(-3, -2, 0, 1, 1, 3)
y <- as.factor(y)
centre(y) # the reference level
centre(y, FUN = raster::modal) # the most frequent (modal) value
library(sf)
data(centro_2015)
centre(centro_2015$padrao) # the reference level
centre(centro_2015$padrao, FUN = raster::modal) # the most frequent value
}
