% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centro_13_15}
\alias{centro_13_15}
\title{Prices of Florianopolis' downtown apartaments in 2013 and 2015}
\format{A tibble with 90 rows and 11 variables:
\itemize{
   \item valor: price, in brazilian Reais
   \item area_total: Total Area, in squared meters
   \item quartos: Rooms
   \item suites: Ensuites
   \item garagens: Garages
   \item dist_b_mar: Distance to the beach
   \item padrao: Building Standard - baixo, medio,
    alto (i.e. low, normal, high)
   \item novo: boolean for brand new apartments -
    sim, nao (TRUE, FALSE)
   \item data_2015: boolean to specify if data is from 2015 -
    sim, nao (TRUE, FALSE)
}}
\source{
\strong{IBAPE/SC}. \emph{Curso de Engenharia de Avaliacoes 
Imobiliarias: modulo basico}. Florianopolis, maio/2015
}
\usage{
centro_13_15
}
\description{
A SpatialPointsDataFrame containing two samples in different dates with 
prices and other attributes in Florianopolis' downtown
}
\keyword{datasets}
