% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGrid.R
\name{createGrid}
\alias{createGrid}
\title{Create grid for predictions}
\usage{
createGrid(x, object, at, ...)
}
\arguments{
\item{object}{object of class lm}

\item{at}{List to be used to calculate the predictions
(defaults for center of each variable).}

\item{\dots}{further arguments passed to predict.lm.}

\item{variable}{variable to be plotted against response variable}

\item{func}{function used to transform the response (optional)}

\item{interval}{the type of interval calculation (provided to predict.lm) to
be ploted.}

\item{level}{Tolerance/confidence level (provided to predict.lm) to
be ploted.}

\item{ca}{(T/F) should the limits of the invertal of arbitration be plotted?}

\item{av}{(T/F) should the arbitrated value be plotted?}
}
\description{
\code{createGrid} create grids to predictions for ploting model variables
(terms) with or without confidence/prediction intervals
}
\examples{
library(sf)
data(centro_2015)
dados <- st_drop_geometry(centro_2015)
mod <- lm(sqrt(valor) ~ sqrt(area_total) + quartos + suites, dados)
p <- createGrid("area_total", mod)
p <- createGrid("area_total", mod,
                at = list(area_total = 205, quartos = 3, suites = 1))
}
