% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ross-heidecke.R
\name{depreciation}
\alias{depreciation}
\alias{linhaReta}
\alias{linhaRetaVariante}
\alias{kuentzle}
\alias{valorDecrescente}
\alias{ross}
\alias{heidecke}
\alias{rossheidecke}
\title{Computes Depreciation parameters according to several methods}
\usage{
linhaReta(VidaUtil, IdadeAparente, ValorResidual = 0)

linhaRetaVariante(IdadeReal, taxa, periodo)

kuentzle(VidaUtil, IdadeAparente)

valorDecrescente(VidaUtil, r)

ross(VidaUtil, IdadeAparente)

heidecke(
  Conservacao = c("Novo", "Entre novo e regular", "Regular",
    "Entre regular e reparos simples", "Reparos simples",
    "Entre reparos simples e importantes", "Reparos importantes",
    "Entre reparos importantes e sem valor", "Sem valor")
)

rossheidecke(
  VidaUtil,
  IdadeAparente,
  ValorResidual = 0,
  Conservacao = c("Novo", "Entre novo e regular", "Regular",
    "Entre regular e reparos simples", "Reparos simples",
    "Entre reparos simples e importantes", "Reparos importantes",
    "Entre reparos importantes e sem valor", "Sem valor")
)
}
\arguments{
\item{VidaUtil}{lifespan}

\item{IdadeAparente}{apparent age}

\item{ValorResidual}{residual value, in percentage}

\item{r}{depreciation rate}

\item{Conservacao}{Conservation state}
}
\value{
The Ross depreciation parameter

The Heidecke depreciation parameter

The Ross-Heidecke depreciation parameter
}
\description{
Computes Depreciation parameters according to several methods
}
\examples{
linhaReta(60, 25, .20)

linhaRetaVariante(0, .07, 5)
linhaRetaVariante(5, .07, 5)
linhaRetaVariante(10, .07, 5)
linhaRetaVariante(15, .07, 5)
linhaRetaVariante(20, .07, 5)
linhaRetaVariante(25, .07, 5)

# Residencias proletario rustico a medio comercial
valorDecrescente(25, r = 0.015)
valorDecrescente(67, r = 0.015)

# Residencias medio superior a luxo
valorDecrescente(25, r = 0.02)
valorDecrescente(50, r = 0.02)
valorDecrescente(50, r = 0.10)

ross(50, 10)

heidecke(Conservacao = "Novo")
heidecke("Entre novo e regular")
heidecke("Regular")
heidecke("Entre regular e reparos simples")
heidecke("Reparos simples")
heidecke("Entre reparos simples e importantes")
heidecke("Reparos importantes")
heidecke("Entre reparos importantes e sem valor")
heidecke("Sem valor")

rossheidecke(30, 0, Conservacao = "Novo")
rossheidecke(30, 10,  Conservacao = "N")
rossheidecke(30, 20,  Conservacao = "N")
rossheidecke(30, 30,  Conservacao = "N")

rossheidecke(30, 0, Conservacao = "Reparos simples")
rossheidecke(30, 10, Conservacao = "Reparos simples")
rossheidecke(30, 20, Conservacao = "Reparos simples")
rossheidecke(30, 30, Conservacao = "Reparos simples")

rossheidecke(50, 20, Conservacao = "Reparos importantes")
rossheidecke(50, 20, .18, "Reparos importantes")

rossheidecke(60, 25, Conservacao = "Reparos importantes")
rossheidecke(60, 25, .20, Conservacao = "Reparos importantes")

# Curva de Depreciacao
plot(NULL, xlim = c(0, 50), ylim = c(0, 1), las = 2,
xlab = "Idade Aparente",
ylab = "Coeficiente de Depreciacao",
main = "Depreciação de um bem")
curve(linhaReta(50, x), from = 1, to = 50, add = T)

curve(kuentzle(50, x), from = 1, to = 50, add = T, col = "green")

curve(valorDecrescente(x, r = .10), from = 1, to = 50, add = T, col ="blue")

curve(ross(50, x), add = T, col = "red")

curve(rossheidecke(50, x, Conservacao = "Regular"),
       add = T, col = "orange")

legend("topright", inset = .02,
      title = "Metodo",
      legend=c("Valor Decrescente", "Linha Reta", "Ross-Heideck (Regular)",
               "Ross", "Kuentzle"), text.font = 4,
      col=c("blue", "black", "orange", "red", "green"), cex=0.8, lty = 1)
}
