% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{equacoes}
\alias{equacoes}
\title{Regression equation}
\usage{
equacoes(object, precision = 10000)
}
\arguments{
\item{object}{object of class \code{lm}}

\item{precision}{the precision required for coefficients}
}
\description{
Givens a \link{lm} object, returns its regression equation
}
\examples{
dados <- st_drop_geometry(centro_2015)
fit <- lm(log(valor) ~ ., dados)
equacoes(fit)
equacoes(fit, precision = 1)
}
