% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{FC}
\alias{FC}
\title{Cash Flow table}
\usage{
FC(fcv, fci, bdi_i, cor, tma)
}
\description{
Cash Flow table
}
\examples{
tma <- 1.91/100
FC(fcv_provavel, fci_provavel, bdi_i = bdi_i, cor = cor, tma = tma)
}
