% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{FCI}
\alias{FCI}
\title{Investment cash flow}
\usage{
FCI(cc, wc, bdi_c, index = seq(0, length(wc) - 1))
}
\arguments{
\item{cc}{construction cost}

\item{wc}{vector of weights of the total constructions costs over periods}

\item{bdi_c}{Constructors Budget Difference Income}

\item{index}{a vector with the periods in which are made the expenses}
}
\value{
a vector containing the expected investment cash flow
}
\description{
Investment cash flow
}
\examples{
wc <- c(5.67, 6.63, 7.24, 7.55, 10.76, 13.26, 14.72, 13.16, 14.18, 6.84)/100
bdi_c <- 31.46/100
cc <-  3523496.76
fci_provavel <- FCI(cc = cc, wc = wc, bdi_c = bdi_c)
}
