% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{FCL}
\alias{FCL}
\title{Net Cash Flow}
\usage{
FCL(fcv, fci, bdi_i, cor)
}
\arguments{
\item{fcv}{the expected sales cash flow}

\item{fci}{the expected investment cash flow}

\item{bdi_i}{Developers Budget Difference Income}

\item{cor}{Brokerage fee}
}
\value{
the expected Net Cash Flow
}
\description{
Net Cash Flow
}
\examples{
bdi_i <- 23.5223/100
cor <- 5/100
fcl_provavel <- FCL(fcv_provavel, fci_provavel, bdi_i = bdi_i, cor = cor)
}
