% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{FCV}
\alias{FCV}
\title{Sales Cash Flow}
\usage{
FCV(vgv, wv, index = seq(0, length(wv) - 1))
}
\arguments{
\item{vgv}{overall selling value}

\item{w}{vector of weights of the total expected sales over periods}
}
\value{
a vector containing the expected sales cash flow
}
\description{
Sales Cash Flow
}
\examples{
wv <- c(0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)/20
fcv_provavel <- FCV(vgv = vgv, wv = wv)
}
