% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros_nbr14653.R
\name{g_precisao}
\alias{g_precisao}
\title{Based on a dataframe containing the estimated values, gives precision degree
parameters based in NBR14.653-2}
\usage{
g_precisao(amplitude)
}
\arguments{
\item{amplitude}{a vector containing the amplitudes of the estimations
\code{\link{bestfit}}}
}
\value{
a vector containing the precision degrees of the estimations.
}
\description{
Precision degree based on NBR14.653-2.
}
\examples{
fit <- lm(log(Valor_Total) ~ . - lat - lon, data = centro_2015)
aval <- grepl("^aval", rownames(centro_2015))
Y_hat <- predict(object = fit, interval = "confidence", newdata = centro_2015[aval,])
Y_hat <- inverse(Y_hat, "log")
amp <- amplitude(Y_hat)
g_precisao(amp)

}
