% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_map.R
\name{gen_map}
\alias{gen_map}
\title{Generates map from SpatialPointsDataFrame}
\usage{
gen_map(data,
  proj = "+proj=utm +zone=22 +south +ellps=WGS84 +units=m +no_defs",
  api_key)
}
\arguments{
\item{data}{A SpatialPointsDataFrame.}

\item{proj}{projection definition}
}
\value{
A ggmap plot
}
\description{
Generates map from SpatialPointsDataFrame
}
\examples{

gen_map(centro_2015)
}
