% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glebas}
\alias{glebas}
\title{Urban large parcels data}
\format{
A tibble with 19 rows and 4 variables:
\itemize{
  \item R: id
  \item Ficha: another id (unused)
  \item VU: unitary value per sq. meters
  \item Area: area, in sq. meters
}
}
\usage{
glebas
}
\description{
A tibble containing a sample of 19 large parcels.
}
\examples{
data(glebas)
fit <- lm(rec(VU) ~ log(Area), data = glebas)
library(effects)
plot(predictorEffects(fit, residuals = T), id = T,
      axes = list(
        grid = TRUE,
        y=list(transform=list(trans=rec, inverse=rec), lab = "VU")
        )
     )
# Issue: Influential Point 4 (see also plot(fit))

# Solution:
fit1 <- update(fit, rec(VU)~log(Area), subset = -4)
plot(predictorEffects(fit1, residuals = T), id = T,
      axes = list(
        grid = TRUE,
        y=list(transform=list(trans=rec, inverse=rec), lab = "VU")
      )
    )
}
\keyword{datasets}
