% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glebas3}
\alias{glebas3}
\title{Urban large parcels in different urban contexts}
\format{
A tibble with 17 rows and 5 variables:
\itemize{
  \item R: id
  \item VU: unitary value per sq. meters
  \item AT: land area, in sq. meters
  \item ACESSO: dummy variable that indicates if the area is direct reachble
  or not
  \item SUP: dummy variable that indicates if the area was landfilled
}
}
\usage{
glebas3
}
\description{
A tibble containing a sample of 17 large parcels within differents urban
contexts.
}
\examples{
data(glebas3)
fit <- lm(log(VU) ~ I(AT^-1) + ACESSO + SUP, data = glebas3)
library(effects)
plot(predictorEffects(fit, residuals = T), id = T,
      axes = list(
        grid = TRUE,
        x = list(rotate=30),
        y = list(transform=list(trans=log, inverse=exp), lab = "VU")
))
powerPlot(fit, axis="inverted", smooth = TRUE, methods = c("lm", "loess"))
p <- predict(fit, newdata = list(AT = 60000, ACESSO = factor(0),
                                  SUP = factor(1)),
               interval = "confidence", level = .80
            )
exp(p)
amplitude(exp(p)) # very good!
}
\keyword{datasets}
