% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_summary.R
\name{graphic_summary}
\alias{graphic_summary}
\title{Plot data frame graphical summaries}
\usage{
graphic_summary(formula, data)
}
\arguments{
\item{formula}{a generic formula.}

\item{data}{a dataframe.}
}
\description{
\code{graphical summary} creates standardized boxplots for data frames visualizations.
}
\examples{
data <- as.data.frame(centro_2015@data)
graphic_summary(valor ~ ., data = data)
}
