% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grau.R
\name{grau}
\alias{grau}
\title{Based on a lm object, returns reasoning degree parameters according to
NBR14.653-2.}
\usage{
grau(object)
}
\arguments{
\item{object}{object of class "lm"}
}
\description{
Reasoning degree based on NBR14.653-2.
}
\examples{
fit <- lm(log(valor) ~ ., data = centro_2015@data)
grau(fit)
}
