% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_ratio.R
\name{iaao_Ratio}
\alias{iaao_Ratio}
\alias{iaao_Ratio.lm}
\title{Standard on Ratio Studies function}
\usage{
iaao_Ratio(x, ...)

\method{iaao_Ratio}{lm}(object, func = "identity", ...)
}
\arguments{
\item{\dots}{further arguments passed to \code{iaao_Ratio}.}

\item{object}{An object of class lm.}

\item{func}{function used to transform the response (defaults to identity)}

\item{AssessedValue}{vector of assessed values (valores venais ou valores preditos do modelo)}

\item{SalePrice}{vector of sale prices (preços de mercado ou preços reais do modelo)}
}
\value{
COD and PRD Ratios according to IAOO
}
\description{
Statiscal anlysis for matching of appraisal and market data.
Based on International Association of Assessing Officers' Standard on Ratio Studies Document
}
\examples{
library(sf)
dados <- st_drop_geometry(centro_2015)
dados$padrao <- as.numeric(dados$padrao)
fit <- lm(log(valor)~area_total + quartos + suites + garagens +
log(dist_b_mar) + I(1/padrao), data = dados, subset = -c(31, 39))
iaao_Ratio(object = fit, func = "log")
}
