% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{invFunc}
\alias{invFunc}
\title{Returns the inverse of the function f (character)}
\usage{
invFunc(FUN)
}
\arguments{
\item{FUN}{a function of the box-cox family (rsqr(), rec(), rsqrt(), log(),
cubroot(), sqrt(), I() and sqr())}
}
\value{
the inverse function, in character format.
}
\description{
Returns the inverse of the function f (character)
}
\examples{
invFunc("log")
}
