% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{involutivo_estatico}
\alias{involutivo_estatico}
\title{Real States appraisals based in the involutive method}
\usage{
involutivo_estatico(vgv, cc, bdi_c, bdi_i, cor)
}
\arguments{
\item{vgv}{Overall Sales Value}

\item{cc}{Constructions Costs}

\item{bdi_c}{Constructors Budget Difference Income}

\item{bdi_i}{Developers Budget Difference Income}

\item{cor}{Brokerage fee}
}
\description{
Real States appraisals based in the involutive method
}
