% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{itacorubi_2015}
\alias{itacorubi_2015}
\title{Prices of 63 Florianopolis' apartaments in Itacorubi neighbourhood}
\format{A tibble with 53 rows (50 samples and 3 apartments to be
 appraised) and 9 variables:
\itemize{
   \item valor: price, in brazilian Reais
   \item area_total: Total Area, in squared meters
   \item area_privativa: Private Area, in squared meters
   \item quartos: Rooms
   \item suites: Ensuites
   \item garagens: Garages
   \item padrao: Building Standard - baixo, medio,
    alto (i.e. low, normal, high)
   \item elevador: boolean checking the existance of elevators -
    sim, nao (TRUE, FALSE)
   \item novo: boolean for brand new apartaments -
    novo, usado (TRUE, FALSE)
}}
\source{
\strong{IBAPE/SC}. \emph{Curso de Engenharia de Avaliacoes Imobiliarias: modulo basico}.
Florianopolis, maio/2015
}
\usage{
itacorubi_2015
}
\description{
A SpatialPointsDataFrame containing a sample of 63 apartaments with prices and other
attributes in Florianopolis' Itacorubi neighbourhood
}
\keyword{datasets}
