% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jurere_2017}
\alias{jurere_2017}
\title{Land division data}
\format{
A tibble with 20 rows and 8 variables:
\itemize{
  \item VALOR TOTAL: price, in brazilian Reais
  \item VU: unitary value per sq. meters
  \item AREA: area, in sq. meters
  \item TESTADA: front width of the land in meters
  \item ESQUINA: corner?
  0 = no; 1 = yes
  \item FRENTES: number of lot fronts
  \item DIST_MAR: distance to the sea
  \item PAVIMENTOS: number of floors
  \item DIST_MAR: distance to the sea
  \item CONDOMINIO_FECHADO: gated community?
  0 = no, 1 = yes.
  \item PAVIMENTO: paved?
  NAO = no, SIM = yes.
  \item TOPOGRAFIA: topography
  PLANO = flat
  \item DATA: date
  \item FONTE: source
}
}
\usage{
jurere_2017
}
\description{
A tibble containing a sample of 20 plots in subdivision in Florianopolis.
Paradigm situation: dry, flat, 15m front width and 30~60m length.
}
\examples{
data(jurere_2017)
jurere_2017$PAVIMENTOS <- as.factor(jurere_2017$PAVIMENTOS)
fit <- lm(VU ~ AREA + TESTADA + log(DIST_MAR) + PAVIMENTOS,
          data = jurere_2017, subset = -c(11, 27, 29))
}
\keyword{datasets}
