% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxAndMins.R
\name{MaxsAndMins}
\alias{MaxsAndMins}
\title{MaxsAndMins}
\usage{
MaxsAndMins(newdata, data)
}
\arguments{
\item{newdata}{a tibble with new data, *i.e.* prediction data}

\item{data}{a data frame}
}
\value{
a data frame with the newdata values above max or below min original data
}
\description{
Find out how much new data extrapolates the model data
}
\examples{
dados <- st_drop_geometry(centro_2015)
fit <- lm(log(valor) ~ ., data = dados)
newdata <- new_data(fit)
MaxsAndMins(newdata, dados)
newdata[1, "area_total"] <- 600
newdata[2, "dist_b_mar"] <- 30
MaxsAndMins(newdata, dados)

}
