% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_analysis.R
\name{outlier_analysis}
\alias{outlier_analysis}
\title{Sample outlier analysis}
\usage{
outlier_analysis(x, criterion = c("30_percent", "2_sd", "chauvenet"))
}
\arguments{
\item{x}{a vector of class numeric.}

\item{criterion}{criterion to be used.}
}
\value{
data
}
\description{
Performs sample outlier analysis according to chosen criteria.
}
\examples{
outlier_analysis(x = trindade, criterion = "30_percent")
outlier_analysis(x = trindade, criterion = "2_sd")
outlier_analysis(x = trindade, criterion = "chauvenet")
}
