% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros_nbr14653.R
\name{amplitude}
\alias{amplitude}
\alias{g_precisao}
\alias{campo_arbitrio}
\alias{aval_intervalar}
\title{Based on a dataframe containing the estimated values, returns amplitude
according to NBR14.653-2}
\usage{
amplitude(Y)

g_precisao(amplitude)

campo_arbitrio(Y)

aval_intervalar(Y)
}
\arguments{
\item{Y}{dataframe with \emph{Y_hat} values with confidence/prediction
\code{\link{bestfit}}}

\item{amplitude}{a vector containing the amplitudes of the estimations
\code{\link{bestfit}}}

\item{Y}{dataframe with \emph{Y_hat} values with or without
confidence/prediction \code{\link{bestfit}}}

\item{Y}{dataframe with \emph{Y_hat} values with or without confidence/prediction 
\code{\link{bestfit}}}
}
\value{
a vector containing the amplitudes of the \emph{Y_hat} in \%

a vector containing the precision degrees of the estimations.

a vector containing the interval of arbitration according to
 NBR14.653-2.

a vector containing interval evaluation according to NBR14.653-2.
}
\description{
Amplitude according to NBR14.653-2.

Precision degree based on NBR14.653-2.

Interval of arbitration according to NBR14.653-2.

Interval evaluation based on NBR14.653-2.
}
\examples{
fit <- lm(log(valor) ~ ., data = centro_2015@data)
aval <- centro_2015@data \%>\% filter(is.na(valor))
Y_hat <- predict(object = fit, interval = "confidence", newdata = aval)
Y_hat <- inverse(Y_hat, "log")
amp <- amplitude(Y_hat)
amp
g_precisao(amp)
campo_arbitrio(Y_hat)
aval_intervalar(Y_hat)
}
