% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdf.R
\name{plotdf}
\alias{plotdf}
\title{Plot data frames}
\usage{
plotdf(formula, data)
}
\arguments{
\item{formula}{a generic formula.}

\item{data}{a data frame.}
}
\description{
\code{plotdf} creates standardized plots for easy data frames visualizations.
}
\examples{
dados <- st_drop_geometry(centro_2015)
plotdf(valor ~ ., dados)
plotdf(log(valor) ~ ., dados)
plotdf(log(valor) ~ area_total + quartos + suites + garagens +
log(dist_b_mar) + padrao , dados)
}
