% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVar.R
\name{plotFactor}
\alias{plotFactor}
\title{export}
\usage{
plotFactor(
  x,
  object,
  interval = c("none", "confidence", "prediction", "both"),
  level = 0.8,
  local,
  func,
  ca = FALSE,
  av,
  elasticidade = TRUE,
  ...
)
}
\description{
export
}
