% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotModel}
\alias{plotModel}
\alias{plotModel.bestfit}
\alias{plotModel.lm}
\title{Plot bestfit models}
\usage{
plotModel(object, ...)

\method{plotModel}{bestfit}(object, fit = 1, ...)

\method{plotModel}{lm}(object, func = "identity", ...)
}
\arguments{
\item{object}{Object of class \code{\link{bestfit}}}

\item{\dots}{further arguments passed to predict.lm.}

\item{fit}{the number of the chosen fit from the combination matrix (defaults
for the best fit found with \code{\link{bestfit}}).}

\item{func}{function used to transform the response (optional)}

\item{interval}{the type of interval calculation (provided to predict.lm) to
be ploted with the model.}

\item{level}{Tolerance/confidence level (provided to predict.lm) to be
ploted.}

\item{local}{Data Frame to be used for calculate the estimates
(defaults for center of each variable).}
}
\description{
\code{plotModel} plots \code{\link{bestfit}} models
}
\examples{
data <- st_drop_geometry(centro_2015)
best_fit <- bestfit(valor ~ .,  data = data)
plotModel(best_fit, interval = "confidence")
plotModel(best_fit, fit = 2, interval = "confidence")
plotModel(best_fit, interval = "confidence",
        local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"))

fit <- lm(log(valor) ~ ., data = data)
plotModel(fit)
plotModel(fit, interval = "confidence")
plotModel(fit, interval = "confidence", func = "log")
plotModel(fit, interval = "confidence",
        local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"))
plotModel(fit, interval = "confidence", func = "log",
        local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"))

}
