% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{porcento}
\alias{porcento}
\alias{pct}
\title{Write in percentage form}
\usage{
porcento(x, ...)

pct(x, ...)
}
\arguments{
\item{x}{a real number}
}
\description{
This function transforms any number to percentage format for reporting purposes.
}
\examples{
porcento(0.25)
pct(0.25)
}
