% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bestfit.R
\name{predict.bestfit}
\alias{predict.bestfit}
\title{Predict method for bestfit linear models}
\usage{
\method{predict}{bestfit}(object, fit = 1, newdata = object$newdata,
  interval = c("none", "confidence", "prediction"), level = 0.8, ...)
}
\arguments{
\item{object}{object of class \code{\link{bestfit}}}

\item{fit}{The number of the chosen model from the combinations matrix
(defaults for the best R2 model)}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. Defaults for the eventually existent "aval" data in data
frame used to build the \code{bestfit} model. If omitted, the fitted values
are used.}

\item{interval}{Type of interval calculation. Can be abbreviated. Defaults to
confidence interval}

\item{level}{Tolerance/confidence level. Defaults for the standardized value
in NBR-14.653-2 (80\%)}

\item{\dots}{further arguments passed to predict.lm.}
}
\description{
Predict values based on a \code{\link{bestfit}} object.
}
\examples{
best_fit <- bestfit(valor ~ ., data = centro_2015@data)
p <- predict(best_fit, interval = "confidence")
p

predict(best_fit, fit = 2, interval = "confidence")

}
