% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictResponse.R
\name{predictResponse}
\alias{predictResponse}
\title{Mass grid predictions for marginal effects plotting}
\usage{
predictResponse(
  x,
  object,
  interval = c("none", "confidence", "prediction", "both"),
  level = 0.8,
  at,
  FUN,
  residuals = FALSE,
  ...
)
}
\description{
\code{predictResponse} makes predictions for plotting marginal effects for
model terms with or without confidence/prediction intervals
}
\examples{
library(sf)
data(centro_2015)
centro_2015 <- within(centro_2015, VU <- valor/area_total)
fit<- lm(log(VU) ~ log(area_total) + quartos + suites +
          garagens + log(dist_b_mar) + padrao,
          data = centro_2015, subset = -c(31, 39))
p <- predictResponse("area_total", fit)
p
}
