% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Reais}
\alias{Reais}
\alias{reais}
\title{Wrapper around brformat}
\usage{
Reais(x, prefix = "R$", ...)

reais(...)
}
\arguments{
\item{x}{a real number}

\item{prefix}{currency units. Defaults for brazilian reais.}

\item{\ldots}{further arguments to be passed to \link{brformat}.}
}
\value{
text in Reais.
}
\description{
This is a wrapper around \link{brformat}.
}
\examples{
Reais(100)
library(ggplot2)
p <- ggplot(centro_2015@data, aes(x = area_total, y = valor)) + geom_point()
suppressWarnings(print(p + scale_y_continuous(labels = reais(nsmall = 0))))
}
