% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renda}
\alias{renda}
\title{Income}
\format{a tibble with 7 rows and 5 variables:
\itemize{
  \item area_total: Total Area, in squared meters
  \item valor_aluguel: Rental values
  \item valor_venda: Sale Value
  \item vacancia: Vacancy
  \item condominio: condominium fee
}}
\source{
\strong{HOCHHEIM, Norberto}. \emph{Engenharia de Avaliacoes I}. Florianopolis: IBAPE/SC, 2005, p.99
}
\usage{
renda
}
\description{
A tibble containing a sample of 7 apartments to be used to appraise by the the Income Method.
}
\keyword{datasets}
