% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alugueis.R
\name{rent}
\alias{rent}
\title{Rent calculation}
\usage{
rent(qt, qb, S, A, FC, Fd, CA, it, ib)
}
\arguments{
\item{qt}{lot value per square meter}

\item{qb}{built-up value per square meter}

\item{S}{lot Area}

\item{A}{built-up area}

\item{FC}{factor to be applied to land and built-areas values in order to
obtain the real estate value}

\item{Fd}{depreciation multiplicator}

\item{CA}{Land Use Coefficient}

\item{it}{capitalization rate to be applied to the land value}

\item{ib}{capitalization rate to be applied to the built-up area value}
}
\value{
The rent value for the real estate
}
\description{
Computes the rent of an entire building
}
\examples{
rent(qt = 2000, qb = 1200, S = 400, A = 450, FC = 1.1, Fd = 0.8, CA = 0.9375,
     it = 8/100, ib = 8/100)
}
