% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ross-heidecke.R
\name{rossheidecke}
\alias{rossheidecke}
\title{Computes the depreciation parameter according to the Ross-Heidecke method}
\usage{
rossheidecke(VU, IA, Conservacao = c("Novo", "Entre novo e regular",
  "Regular", "Entre regular e reparos simples", "Reparos simples",
  "Entre reparos simples e importantes", "Reparos importantes",
  "Entre reparos importantes e sem valor", "Sem valor"))
}
\arguments{
\item{VU}{lifespan}

\item{IA}{apparent age}

\item{Conservacao}{Conservation state}
}
\value{
The Ross-Heidecke depreciation parameter
}
\description{
Computes the depreciation parameter according to the Ross-Heidecke method
}
\examples{
rossheidecke(30, 0, "Novo")
rossheidecke(30, 10, "N")
rossheidecke(30, 20, "N")
rossheidecke(30, 30, "N")

rossheidecke(30, 0, "Reparos simples")
rossheidecke(30, 10, "Reparos simples")
rossheidecke(30, 20, "Reparos simples")
rossheidecke(30, 30, "Reparos simples")
}
