% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rsqr}
\alias{rsqr}
\title{Reciprocal of the square of a number}
\usage{
rsqr(x)
}
\arguments{
\item{x}{a numeric vector or array}
}
\description{
Reciprocal of the square of a number
}
\examples{

rsqr(2)
rsqr(1:10)
}
