% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shinyApp}
\usage{
runExample()
}
\description{
\code{runExample} starts a shinyApp that uses main functions in \code{REATOOLS} package
}
\examples{
\dontrun{
REATOOLS::runExample()
}

}
