% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{sensibilidade_bdi_c}
\alias{sensibilidade_bdi_c}
\title{Sensibilidade-BDI do Construtor}
\usage{
sensibilidade_bdi_c(range, cc, wc, vgv, wv, bdi_i, bdi_c, cor, tma)
}
\description{
Sensibilidade-BDI do Construtor
}
\examples{
range_bdi_c <- c(0.9, 1.1)
s_bdi_c <- sensibilidade_bdi_c(range = range_bdi_c,
                               cc = cc, wc = wc,
                               vgv = vgv, wv = wv,
                               bdi_i = bdi_i, bdi_c = bdi_c,
                               cor = cor, tma = tma
                               )
s_bdi_c
}
