% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{sensibilidade_bdi_i}
\alias{sensibilidade_bdi_i}
\title{Sensibilidade-BDI do Incorporador}
\usage{
sensibilidade_bdi_i(range, cc, wc, vgv, wv, bdi_i, bdi_c, cor, tma)
}
\description{
Sensibilidade-BDI do Incorporador
}
\examples{
range_bdi_i <- c(0.9, 1.1)

s_bdi_i <- sensibilidade_bdi_i(range_bdi_i,
                             cc = cc, wc = wc,
                             vgv = vgv, wv = wv,
                             bdi_i = bdi_i, bdi_c = bdi_c,
                             cor = cor, tma = tma)
s_bdi_i
}
