% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{sensibilidade_custo}
\alias{sensibilidade_custo}
\title{Sensibilidade-custo do VPL}
\usage{
sensibilidade_custo(range, cc, wc, vgv, wv, bdi_i, bdi_c, cor, tma)
}
\description{
Sensibilidade-custo do VPL
}
\examples{
range_custos <- c(.9, 1.1)
s_custo <- sensibilidade_custo(range = range_custos,
                               cc = cc, wc = wc,
                               vgv = vgv, wv = wv,
                               bdi_i = bdi_i, bdi_c = bdi_c,
                               cor = cor, tma = tma)
s_custo

ggplot(s_custo, aes(x = CC, y = VPL)) +
  geom_line() + geom_point() +
  scale_x_continuous(labels = scales::dollar) +
  scale_y_continuous(labels = scales::dollar) +
  labs(title = "Sensibilidade à variação dos Custos de Construção")
}
