% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{sensibilidade_tma}
\alias{sensibilidade_tma}
\title{Sensibilidade aa TMA}
\usage{
sensibilidade_tma(range, fcl)
}
\description{
Sensibilidade aa TMA
}
\examples{
range_tma <- c(1.2, 2.6)/100
s_tma <- sensibilidade_tma(range_tma, fcl_provavel)
s_tma

require(ggplot2)
ggplot(s_tma, aes(x = TMA, y = VPL)) +
  geom_line() + geom_point() +
  scale_x_continuous(labels = scales::percent) +
  scale_y_continuous(labels = scales::dollar) +
  labs(title = "Sensibilidade à TMA")
}
