% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{sensibilidade_venda}
\alias{sensibilidade_venda}
\title{Sensibilidade-venda do VPL}
\usage{
sensibilidade_venda(range, cc, wc, vgv, wv, bdi_i, bdi_c, cor, tma)
}
\description{
Sensibilidade-venda do VPL
}
\examples{
range_vgv <- c(.9, 1.1)
s_vgv <- sensibilidade_venda(range = range_vgv,
                               cc = cc, wc = wc,
                               vgv = vgv, wv = wv,
                               bdi_i = bdi_i, bdi_c = bdi_c,
                               cor = cor, tma = tma)
s_vgv
ggplot(s_venda, aes(x = Vendas, y = VPL)) +
  geom_point() + geom_line() +
  scale_x_continuous(labels = scales::dollar) +
  scale_y_continuous(labels = scales::dollar) +
  labs(title = "Sensibilidade à variação dos Preços de Venda")
}
