% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkPlot.R
\name{shrinkPlot}
\alias{shrinkPlot}
\title{Create plots to show prediction shrinkage}
\usage{
shrinkPlot(object, FUN = "identity", ...)
}
\arguments{
\item{object}{object of class \code{\link{bestfit}}}

\item{FUN}{function used to transform the response (defaults to identity)}
}
\value{
a shrinkage prediction plot
}
\description{
Computes fitted values from lm package and plots aside of observed values
}
\examples{
library(sf)
dados <- st_drop_geometry(centro_2015)
fit0 <- lm(valor ~ 1, dados)
powerPlot(fit0, axis = "inverted")
shrinkPlot(fit0)
dados$padrao <- as.numeric(dados$padrao)
fit <- lm(log(valor)~area_total + quartos + suites + garagens +
log(dist_b_mar) + I(1/padrao), data = dados, subset = -c(31, 39))
shrinkPlot(fit)
shrinkPlot(fit, FUN = "log")
}
