% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bestfit.R
\name{summary.bestfit}
\alias{summary.bestfit}
\title{Summary method for class bestfit}
\usage{
\method{summary}{bestfit}(object, fit = 1, subset, ...)
}
\arguments{
\item{object}{an object of class \code{\link{bestfit}}.}

\item{fit}{the number of the chosen fit from the combination matrix (defaults
for the best fit found with \code{\link{bestfit}}).}

\item{subset}{a specification of the rows to be used: defaults to all rows.
This can be any valid indexing vector (see \link{[.data.frame}) for the
rows of data or if that is not supplied, a data frame made up of the
variables used in \code{formula}.}

\item{\dots}{further arguments passed to \link{lm}.}
}
\value{
Returns the call for the \code{\link{bestfit}} function, the best
  (chosen) fit number, the \code{lm }formula and the \code{lm} fit summary
  for the best (chosen) fit transformations found by \code{\link{bestfit}}.
}
\description{
\code{summary} method for class \code{\link{bestfit}} produce result summary
containing the formula for the best (chosen) fit and the \code{summary.lm}
for that fit.
}
\examples{
best_fit <- bestfit(valor ~ ., data = centro_2015@data)
summary(best_fit)
summary(best_fit, fit = 2)
}
