% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terrenos_praia}
\alias{terrenos_praia}
\title{Prices of 48 plots at the beach}
\format{a tibble with 48 rows and 8 columns
\itemize{
  \item valor_unitario unitary value per squared meter.
  \item area area in squared meters.
  \item testada plot front in meters.
  \item profundidade land deep in meters.
  \item pedologia pedology.
  \item frente_mar front of the beach (yes (1) or no (0))
  \item dist_b_mar distance to the sea.
  \item zoneamento zone.
}}
\source{
\strong{HOCHHEIM, Norberto}. \emph{Engenharia de Avaliacoes II}. Florianopolis: IBAPE/SC, 2005, p.E2.2
}
\usage{
terrenos_praia
}
\description{
A tibble containing a sample of 48 unitary values of plots sales prices per squared meters.
}
\keyword{datasets}
