% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trindade}
\alias{trindade}
\title{Prices of 15 Florianopolis' apartaments with 2 rooms in Trindade neighbourhood}
\format{a tibble with 15 rows and 1 variable:
\itemize{
  \item VU unitary value per squared meter.
}}
\source{
\strong{HOCHHEIM, Norberto}. \emph{Engenharia de Avaliacoes I}. Florianopolis: IBAPE/SC, 2005, p.18
}
\usage{
trindade
}
\description{
A tibble containing a sample of 15 unitary values of apartment sales prices per squared meter.
}
\keyword{datasets}
