% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trivelloni_2005}
\alias{trivelloni_2005}
\title{Prices of 249 Florianopolis' Real Estate}
\format{A tibble with 249 rows and 7 variables:
\itemize{
   \item valor: price, in brazilian Reais
   \item tipo: type of property -
    casa, apartamento, kitinete, terreno, comercial
    (i.e, house, apartament, kitchenette, land, commercial)
   \item area_total: Total Area, in squared meters
   \item area_terreno: Ground Area, in squared meters (houses only)
   \item garagens: Garages
   \item novo: boolean for brand new apartments -
    novo, usado (TRUE, FALSE)
   \item padrao: Building Standard -
    baixo, medio, alto (i.e. low, normal, high)
}}
\source{
\strong{Trivelloni, Carlos Alberto Peruzzo}. 
\emph{Metodo para determinacao do valor da localizacao com uso de tecnicas 
inferenciais e geoestatisticas na avaliacao em massa de imoveis}. 
Florianopolis, 2005.
}
\usage{
trivelloni_2005
}
\description{
A SpatialPointsDataFrame containing a sample of 249 Real Estate with prices and other
attributes in Florianopolis' continental area
}
\keyword{datasets}
