% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.bestfit.R
\name{update.bestfit}
\alias{update.bestfit}
\title{Update method for bestfit}
\usage{
\method{update}{bestfit}(object, formula., ...)
}
\arguments{
\item{object}{a bestfit object.}

\item{formula.}{a generic formula.}

\item{\dots}{further arguments to be added to the original call.}
}
\value{
a new bestfit object, update according to the arguments passed to \code{update.bestfit}.
}
\description{
This functions implements the \code{update} method for \link{bestfit} models.
}
\examples{

best_fit <- bestfit(Valor_Total ~ . - lat - lon, centro_2015)
s <- summary(best_fit)
out <- car::outlierTest(s$fit)

#update best_fit in order to exclude the outlier found (AP_31)
outliers <- match(names(out$p), rownames(centro_2015))
update(best_fit, subset = -outliers)

}
