% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{write.kml}
\alias{write.kml}
\title{Title}
\usage{
write.kml(spobj, dsn, layer, var.name, col = bpy.colors(20))
}
\arguments{
\item{spobj}{}

\item{dsn}{}

\item{layer}{}

\item{var.name}{}

\item{col}{}
}
\description{
Title
}
