% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zilli_2020}
\alias{zilli_2020}
\title{Prices of 225 Florianopolis' apartaments in 3 neighborhoods}
\format{
A tibble with 53 rows (50 samples and 3 apartments to be
  appraised) and 7 variables:
\itemize{
  \item VT: price, in brazilian Reais
  \item VU: price per sq. meter
  \item AP: Private Area, in squared meters
  \item DPXV: Distance to Praça XV
  \item DSBM: Distance to Beira Mar Mall
  \item DSIG: Distance to Iguatemi Mall
  \item DCTC: Distance to  CTC/UFSC
  \item DABM: Distance to Beira Mar Avenue
  \item ND: Number of rooms
  \item NB: Number of bathrooms
  \item NS: Number of ensuites
  \item NG: Number of garages
  \item MO: Furnishes - N (none), SM (some), MO (full)
  \item PSN: Swimming pool?
  \item CH: Barbecue grill?
  \item PC: Building Standard - B, M, A
  (i.e. low, normal, high)
  \item BRO: Neighborhood
}
}
\source{
\strong{ZILLI, Carlos Augusto}. \emph{Regressão geograficamente
ponderada aplicada na avaliação em massa de imóveis urbanos.}. 2020.
Dissertação de Mestrado em Engenharia de Transportes e Gestão Territorial.
Centro Tecnológico da UFSC. Florianópolis/SC.
}
\usage{
zilli_2020
}
\description{
A SpatialPointsDataFrame containing a sample of 50 apartaments with
prices and other attributes in Florianopolis' downtown
}
\examples{
data(zilli_2020)
zilli_2020$PC <- as.numeric(zilli_2020$PC)
fit <- lm(log(VU) ~ log(AP) + log(DABM) + ND + NB + NG + PSN + PC,
data = zilli_2020[1:190, ], subset = -c(86, 115))
summary(fit)

fefit <- lm(log(VU) ~ log(AP) + log(DABM) + ND + NB + NG + PSN + PC + BRO,
data = zilli_2020[1:190, ], subset = -c(86, 115))
summary(fefit)

}
\keyword{datasets}
