% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxplot.R
\name{bboxplot}
\alias{bboxplot}
\alias{bboxplot.default}
\alias{bboxplot.formula}
\title{Beautiful boxplots with ggplot2.}
\usage{
bboxplot(y, ...)

\method{bboxplot}{default}(y, g, data, ...)

\method{bboxplot}{formula}(formula, data, ...)
}
\arguments{
\item{y}{a numeric variable for which the boxplot is to be constructed.}

\item{\dots}{further arguments passed to \code{\link[ggplot2]{geom_boxplot}}.}

\item{g}{a grouping variable, usually a factor, for constructing parallel
boxplots.}

\item{data}{a dataframe containing the variables.}

\item{formula}{A model formula of the form ~ y to produce a boxplot for
the variable y, or of the form y ~ g to produce parallel boxplots for y
within levels of the grouping variable(s) g, etc., usually factors.}
}
\value{
A boxplot of the variable \code{y} or the boxplot of the \code{y}
  against \code{g}.
}
\description{
Create boxplots with \link{ggplot2}.
}
\examples{
data <- as.data.frame(centro_2015@data)
bboxplot("valor", data = data)
bboxplot("valor", "padrao", data)
bboxplot(valor~padrao, data)
bboxplot(~valor, data)
}
