% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{brformat}
\alias{brformat}
\title{Wrapper around format}
\usage{
brformat(x, decimal.mark = ",", big.mark = ".", digits = 2,
  nsmall = 2, scientific = FALSE, ...)
}
\arguments{
\item{x}{a number to be formatted by \code{format}}
}
\description{
This function is only a wrapper around \link{format} function that uses standard 
brazillian formats by default
}
