% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centro_2015}
\alias{centro_2015}
\title{Prices of 50 Florianopolis' downtown apartaments}
\format{A tibble with 53 rows (50 samples and 3 apartments to be
  appraised) and 7 variables:
\itemize{
  \item valor: price, in brazilian Reais
  \item area_total: Total Area, in squared meters
  \item quartos: Rooms
  \item suites: Ensuites
  \item garagens: Garages
  \item dist_b_mar: Distance to the beach
  \item padrao: Building Standard - baixo, medio, alto
  (i.e. low, normal, high)
}}
\source{
\strong{HOCHHEIM, Norberto}. \emph{Engenharia de avaliacoes 
imobiliarias: Modulo Basico}. Florianopolis: IBAPE/SC, 2015, p.21-22
}
\usage{
centro_2015
}
\description{
A SpatialPointsDataFrame containing a sample of 50 apartaments with 
prices and other attributes in Florianopolis' downtown
}
\keyword{datasets}
