% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGrid.R
\name{createGrid}
\alias{createGrid}
\title{Create grid for predictions}
\usage{
createGrid(x, object, at, ...)
}
\arguments{
\item{object}{object of class lm}

\item{at}{List to be used to calculate the predictions
(defaults for center of each variable).}

\item{\dots}{further arguments passed to predict.lm.}

\item{variable}{variable to be plotted against response variable}

\item{func}{function used to transform the response (optional)}

\item{interval}{the type of interval calculation (provided to predict.lm) to
be ploted.}

\item{level}{Tolerance/confidence level (provided to predict.lm) to
be ploted.}

\item{ca}{(T/F) should the limits of the invertal of arbitration be plotted?}

\item{av}{(T/F) should the arbitrated value be plotted?}
}
\description{
\code{createGrid} create grids to predictions for ploting model variables
(terms) with or without confidence/prediction intervals
}
\examples{
# Crete random bivariate normal data just for testing
library(MASS)
sample_mean <- c(10000, 250)
sample_cov <- matrix(c(1000^2, -37500,
                       -37500, 50^2),
                     ncol = 2, byrow = T)
n <- 10
set.seed(1)
dados <- mvrnorm(n = n,
                mu = sample_mean,
                Sigma = sample_cov)
colnames(dados) <- c("VU", "Area")
dados <- as.data.frame(dados)

fit <- lm(VU ~ Area, data = dados)
createGrid("Area", fit)
#
# Test with real data
library(sf)
data(centro_2015)
centro_2015 <- within(centro_2015, VU <- valor/area_total)
fit <- lm(sqrt(VU) ~ sqrt(area_total) + quartos + suites + garagens +
          sqrt(dist_b_mar) + padrao,
          data = centro_2015)
createGrid("area_total", fit)
createGrid("area_total", fit,
           at = list(area_total = 205, quartos = 3, suites = 1,
                     garagens = 2, dist_b_mar = 250, padrao = "medio"))

}
