% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticidade.R
\name{elasticidade}
\alias{elasticidade}
\title{Elasticity calculation function}
\usage{
elasticidade(object, variable, func = "log", local, fator = 0.1)
}
\arguments{
\item{object}{object of class "lm"}

\item{variable}{regressor to calculate elasticity}

\item{func}{}
}
\description{
Based on a lm object, returns elasticity based in any parameter in the model.
}
\examples{
data <- centro_2015@data
fit <- lm(log(valor) ~ ., data = data)
elasticidade(fit, "area_total", 
             local = list(area_total = 205, quartos = 3, suites = 1, 
             garagens = 2, dist_b_mar = 250, padrao = "medio"))
elasticidade(fit, "dist_b_mar", 
             local = list(area_total = 205, quartos = 3, suites = 1, 
             garagens = 2, dist_b_mar = 250, padrao = "medio"))
elasticidade(fit, "area_total", 
             local = list(area_total = 205, quartos = 3, suites = 1, 
             garagens = 2, dist_b_mar = 250, padrao = "medio"),
             fator = -0.1)
elasticidade(fit, "padrao", 
             local = list(area_total = 205, quartos = 3, suites = 1, 
             garagens = 2, dist_b_mar = 250, padrao = "medio"),
             fator = +1) 
elasticidade(fit, "padrao", 
             local = list(area_total = 205, quartos = 3, suites = 1, 
             garagens = 2, dist_b_mar = 250, padrao = "medio"),
             fator = -1)              
}
