% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{equacoes}
\alias{equacoes}
\title{Regression equation}
\usage{
equacoes(object)
}
\arguments{
\item{object}{object of class \code{lm}}
}
\description{
Givens a \link{lm} object, returns its regression equation
}
\examples{
dados <- centro_2015@data
fit <- lm(log(valor) ~ ., dados)
equacoes(fit)
}
