% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{equacoes}
\alias{equacoes}
\title{Regression equation}
\usage{
equacoes(
  object,
  type = c("reg", "est"),
  inline = TRUE,
  func,
  accuracy = 100,
  f = round,
  errorTerm = TRUE
)
}
\arguments{
\item{object}{object of class \code{lm}}

\item{type}{the equation type required: regression (reg) or estimation (est).}

\item{inline}{the equation mode. TRUE for inline equations or FALSE for
displayed mode.}

\item{func}{transformation applied to dependent variable.}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: floor, ceiling or round}
}
\description{
Givens a \link{lm} object, returns its regression equation
}
\examples{
dados <- st_drop_geometry(centro_2015)
fit <- lm(log(valor) ~ ., dados)
equacoes(fit)
equacoes(fit, precision = 1)
}
