% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{equacoes}
\alias{equacoes}
\title{Regression equation}
\usage{
equacoes(
  object,
  type = c("reg", "est"),
  inline = TRUE,
  FUN,
  accuracy = 100,
  f = round,
  errorTerm = TRUE
)
}
\arguments{
\item{object}{object of class \code{lm}}

\item{type}{the equation type required: regression (reg) or estimation (est).}

\item{inline}{the equation mode. TRUE for inline equations or FALSE for
displayed mode.}

\item{FUN}{transformation applied to dependent variable.}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: floor, ceiling or round}
}
\description{
Givens a \link{lm} object, returns its regression equation
}
\examples{
data(centro_2015)
centro_2015 <- within(centro_2015, VU <- valor/area_total)
fit <- lm(log(VU) ~ log(area_total) + quartos + suites + garagens +
           log(dist_b_mar) + padrao, centro_2015)
equacoes(fit)
equacoes(fit, accuracy = .01)
}
