% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.R
\name{extrapolate}
\alias{extrapolate}
\title{Extrapolation limits}
\usage{
extrapolate(object, newdata)
}
\arguments{
\item{object}{object of type bestfit}

\item{newdata}{data frame with the real estate data to be appraised}
}
\description{
\code{extrapolate} evaluates the extrapolation eventually used in the
appraisals and returns the precision degree as defined in brazilian standard
NBR-14.653-2 .
}
\examples{
## newdata inside the data file
data(centro_2015)
fit <- lm(log(valor) ~ ., centro_2015@data)
extrapolate(object = fit)

## newdata provided as an argument
library(dplyr)
newdata <- 
  centro_2015@data \%>\% 
  filter(is.na(valor)) \%>\%
  select(-valor)
newdata$area_total[1] <- 1.5*max(centro_2015@data[complete.cases(centro_2015@data), "area_total"])
newdata$dist_b_mar[2] <- .5*min(centro_2015@data[complete.cases(centro_2015@data), "dist_b_mar"])

extrapolate(object = fit, newdata = newdata)
}
