% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_summary.R
\name{graphic_summary}
\alias{graphic_summary}
\title{Plot data frame graphical summaries}
\usage{
graphic_summary(formula, df)
}
\arguments{
\item{formula}{a generic formula.}

\item{df}{a dataframe.}
}
\description{
\code{graphical summary} creates standardized boxplots for data frames visualizations.
}
\examples{
graphic_summary(valor ~ ., centro_2015@data)
}
