% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_ratio.R
\name{iaao_Ratio}
\alias{iaao_Ratio}
\title{Standard on Ratio Studies function}
\usage{
iaao_Ratio(AssessedValue, SalePrice, OutlierTrimming = FALSE)
}
\arguments{
\item{AssessedValue}{vector of assessed values (valores venais ou valores preditos do modelo)}

\item{SalePrice}{vector of sale prices (preços de mercado ou preços reais do modelo)}

\item{OutlierTrimming}{boolean for outliers trimming or not}
}
\value{
COD and PRD Ratios according to IAOO
}
\description{
Statiscal anlysis for matching of appraisal and market data.
Based on International Association of Assessing Officers' Standard on Ratio Studies Document
}
\examples{
library(appraiseR)
library(broom)

dados <- centro_2015@data
fit <- lm(log(valor) ~ area_total + quartos + suites + garagens +
log(dist_b_mar) + padrao, data = dados)

dados <- augment(fit)
iaao_Ratio(dados$.fitted, dados$`log(valor)`)
}
