% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_ratio.R
\name{iaao_Ratio}
\alias{iaao_Ratio}
\title{Standard on Ratio Studies function}
\usage{
iaao_Ratio(AssessedValue, SalePrice, OutlierTrimming = FALSE)
}
\arguments{
\item{AssessedValue}{vector of assessed values (valores venais ou valores preditos do modelo)}

\item{SalePrice}{vector of sale prices (preços de mercado ou preços reais do modelo)}

\item{OutlierTrimming}{boolean for outliers trimming or not}
}
\value{
COD and PRD Ratios according to IAOO
}
\description{
Statiscal anlysis for matching of appraisal and market data.
Based on International Association of Assessing Officers' Standard on Ratio Studies Document
}
\examples{
library(appraiseR)

zilli_2020 <- st_drop_geometry(zilli_2020)
zilli_2020$PC <- as.numeric(zilli_2020$PC)
fit <- lm(log(VU) ~ log(AP) + log(DABM) + ND + NB + NG + PSN + PC + BRO,
data = zilli_2020[1:190, ], subset = -c(86, 115))

new = zilli_2020[c(191:213, 215:225), ]
fitted <- predict(fit, newdata = new)
iaao_Ratio(exp(fitted), new[, "VU", drop = TRUE])
}
