% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inverse}
\alias{inverse}
\title{Returns a vector generated with the inverse of the function f}
\usage{
inverse(x, func)
}
\arguments{
\item{x}{A vector or object of type}

\item{func}{a function of the box-cox family (rsqr(), rec(), rsqrt(), log(),
sqrt(), I() and sqr())}
}
\value{
a
}
\description{
Returns a vector generated with the inverse of the function f
}
\examples{
fit <- lm(log(valor) ~ ., data = centro_2015@data)
aval <- new_data(fit)
Y <- predict(fit, newdata = aval, interval = "confidence")
inverse(Y, "log")
}
