% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loteamento}
\alias{loteamento}
\title{Land division data}
\format{A tibble with 20 rows and 8 variables:
\itemize{
  \item valor: price, in brazilian Reais
  \item area: land area in squared meters
  \item tipo: type: offer or sale -
  venda, oferta (i.e. sale, offer)
  \item frente: front width of the land in meters
  \item profundidade: length of the land in meters
  \item topo: topography -
  plano, aclive (i.e. flat, slope)
  \item inclinacao: slope
  \item pedologia: pedology -
  seco, pantanoso (i.e. dry, marshy)
}}
\source{
\strong{HOCHHEIM, Norberto}. \emph{Engenharia de Avaliacoes I}. 
Florianopolis: IBAPE/SC, 2005, p.74
}
\usage{
loteamento
}
\description{
A tibble containing a sample of 20 plots in subdivision in Florianopolis. 
Paradigm situation: dry, flat, 15m front width and 30~60m length.
}
\keyword{datasets}
