% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxsAndMins.R
\name{MaxsAndMins}
\alias{MaxsAndMins}
\title{MaxsAndMins}
\usage{
MaxsAndMins(newdata, data)
}
\arguments{
\item{newdata}{a tibble with newdata1}

\item{maxs}{a vec}

\item{mins}{}
}
\description{
Find out how much new data extrapolates the model data
}
\examples{
dados <- centro_2015@data
newdata <- 
  dados \%>\% 
  filter(is.na(valor)) \%>\%
  select(-valor)
MaxsAndMins(newdata, data)   

  
}
