% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_data}
\alias{new_data}
\title{Builds \code{newdata} argument to be used in \link{predict.lm}}
\usage{
new_data(object)
}
\arguments{
\item{object}{object of class \code{lm}}
}
\description{
Builds a new \code{data.frame} containing only elements
to be appraised from the current \code{data.frame}
}
\examples{
fit <- lm(log(valor) ~ ., data = centro_2015@data)
new_data(fit)
}
