% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parameters}
\alias{parameters}
\alias{parameters.lm}
\alias{parameters.bestfit}
\title{Extract object parameters}
\usage{
parameters(object, ...)

\method{parameters}{lm}(object, ...)

\method{parameters}{bestfit}(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{\dots}{not used.}
}
\value{
the parameters, predictors and response names besides the
original data used to build the model.
}
\description{
Returns the parameters used to build a model.
}
\examples{
fit <- lm(log(valor) ~ ., centro_2015@data)
p <- parameters(fit)
p$parameters
p$predictors
p$response
p$data
best_fit <- bestfit(valor ~ ., centro_2015@data)
parameters(best_fit)
}
