% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmod.R
\name{plotmod}
\alias{plotmod}
\alias{plotmod.bestfit}
\title{Plot bestfit models}
\usage{
plotmod(object, ...)

\method{plotmod}{bestfit}(object, fit = 1, interval = c("none",
  "confidence", "prediction"), level = 0.8, ...)
}
\arguments{
\item{object}{Object of class \code{\link{bestfit}}}

\item{\dots}{further arguments passed to predict.lm.}

\item{fit}{the number of the chosen fit from the combination matrix (defaults
for the best fit found with \code{\link{bestfit}}).}

\item{interval}{the type of interval calculation (provided to predict.lm) to 
be ploted with the model.}

\item{level}{Tolerance/confidence level (provided to predict.lm) to be
ploted.}
}
\description{
\code{plotmod} plots \code{\link{bestfit}} models
}
\examples{
best_fit <- bestfit(valor ~ .,  data = data)
plotmod(best_fit, interval = "confidence")
plotmod(best_fit, fit = 2, interval = "confidence")

}
