% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmod.R
\name{plotvar}
\alias{plotvar}
\title{Plot model predictors against the response variable with intervals}
\usage{
plotvar(object, var, func, interval = c("none", "confidence", "prediction"),
  level = 0.8, ...)
}
\arguments{
\item{object}{object of class lm}

\item{var}{variable to be plotted against response variable}

\item{func}{function used to transform the response (optional)}

\item{interval}{the type of interval calculation (provided to predict.lm) to
be ploted.}

\item{level}{Tolerance/confidence level (provided to predict.lm) to
be ploted.}

\item{\dots}{further arguments passed to predict.lm.}
}
\description{
\code{plotvar} plots model variables with confidence/prediction intervals
}
\examples{
fit <- lm(log(Valor_Total) ~ . - lat - lon, centro_2015)
plotvar(fit, "Area_Total", interval = "confidence")
plotvar(fit, "Area_Total", "log", interval = "confidence")
plotvar(fit, "Padrao", interval = "confidence")
plotvar(fit, "Padrao", "log", interval = "confidence")

}
