% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVar.R
\name{plotVar}
\alias{plotVar}
\title{Plot model predictors against the response variable with intervals}
\usage{
plotVar(
  object,
  variable,
  func,
  interval = c("none", "confidence", "prediction", "both"),
  level = 0.8,
  ca = FALSE,
  av,
  local,
  ...
)
}
\arguments{
\item{object}{object of class lm}

\item{variable}{variable to be plotted against response variable}

\item{func}{function used to transform the response (optional)}

\item{interval}{the type of interval calculation (provided to predict.lm) to
be ploted.}

\item{level}{Tolerance/confidence level (provided to predict.lm) to
be ploted.}

\item{ca}{(T/F) should the limits of the invertal of arbitration be plotted?}

\item{av}{(T/F) should the arbitrated value be plotted?}

\item{local}{Data Frame to be used for calculate the estimates
(defaults for center of each variable).}

\item{\dots}{further arguments passed to predict.lm.}
}
\description{
\code{plotvar} plots model variables with confidence/prediction intervals
}
\examples{
library(sf)
data(centro_2015)
dados <- st_drop_geometry(centro_2015)
fit <- lm(log(valor) ~ ., data = dados)
plotVar(fit, "area_total")
plotVar(fit, "area_total", interval = "confidence")
plotVar(fit, "area_total", interval = "prediction")
plotVar(fit, "area_total", interval = "prediction", ca = TRUE)
plotVar(fit, "area_total", interval = "both")
plotVar(fit, "area_total", "log")
plotVar(fit, "area_total", "log", ca = TRUE)
plotVar(fit, "area_total", "log", interval = "confidence")
plotVar(fit, "area_total", "log", interval = "prediction")
plotVar(fit, "area_total", "log", interval = "both")
plotVar(fit, "area_total", interval = "both", ca = TRUE,
        local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"))
plotVar(fit, "area_total", func = "log", interval = "both", ca = TRUE,
        local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"))
plotVar(fit, "padrao")
plotVar(fit, "padrao", ca = TRUE)
plotVar(fit, "padrao", func = "log", ca = TRUE)
plotVar(fit, "padrao", interval = "confidence")
plotVar(fit, "padrao", interval = "prediction")
plotVar(fit, "padrao", interval = "both")
plotVar(fit, "padrao", func = "log", interval = "confidence",
local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"))
plotVar(fit, "padrao", func = "log", interval = "prediction",
local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"), av = 1100000)
plotVar(fit, "area_total", func = "log", interval = "prediction",
local = list(area_total = 205, quartos = 3, suites = 1, garagens = 2,
        dist_b_mar = 250, padrao = "medio"), av = 1100000, ca = TRUE)
}
