% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPlot.R
\name{powerPlot}
\alias{powerPlot}
\alias{powerPlot.lm}
\alias{powerPlot.bestfit}
\alias{powerPlot.lmerMod}
\title{Power plots with ggplot2}
\usage{
powerPlot(object, ...)

\method{powerPlot}{lm}(object, func, axis = c("standard", "inverted"), ...)

\method{powerPlot}{bestfit}(object, fit = 1, ...)

\method{powerPlot}{lmerMod}(object, func, ...)
}
\arguments{
\item{object}{object of class \code{\link{lm}} or \code{\link{bestfit}}}

\item{\dots}{not used.}

\item{fit}{chosen fit}
}
\value{
a power plot
}
\description{
\code{powerPlot} generates Power Plots for \code{\link{lm}} or
\code{\link{bestfit}} objects with \code{\link{ggplot2}}.
}
\examples{
library(sf)
dados <- st_drop_geometry(centro_2015)
dados$padrao <- as.numeric(dados$padrao)
fit <- lm(log(valor) ~ area_total + quartos + suites + garagens +
log(dist_b_mar) + I(1/padrao), dados, subset = -c(31, 39))
powerPlot(fit)
powerPlot(fit, axis = "inverted")
p <- powerPlot(fit, func = "log", axis = "inverted")
library(ggplot2)
p + labs(title = "Poder de Predição", subtitle = "Em milhões de Reais")
dados <- st_drop_geometry(centro_2015)
best_fit <- bestfit(valor ~ ., dados)
powerPlot(best_fit, fit = 257)
library(lme4)
data(centro_2015)
dados <- st_drop_geometry(centro_2015)
Mfit <- lmer(log(valor) ~ area_total + quartos + suites + garagens +
dist_b_mar + (1|padrao), dados)
powerPlot(Mfit)
powerPlot(Mfit, func = "log")
}
