% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictResponse.R
\name{predictResponse}
\alias{predictResponse}
\title{Mass grid predictions for marginal effects plotting}
\usage{
predictResponse(
  x,
  object,
  interval = c("none", "confidence", "prediction", "both"),
  level = 0.8,
  at,
  func,
  ...
)
}
\description{
\code{predictResponse} makes predictions for plotting marginal effects for
model terms with or without confidence/prediction intervals
}
\examples{
library(sf)
data(centro_2015)
dados <- st_drop_geometry(centro_2015)
mod <- lm(sqrt(valor) ~ sqrt(area_total) + quartos + suites, dados)
p <- predictResponse("area_total", mod)
p
}
