% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reais}
\alias{reais}
\title{Wrapper around brformat}
\usage{
reais(prefix = "R$", ...)
}
\arguments{
\item{prefix}{currency units. Defaults for brazilian reais.}

\item{\ldots}{further arguments to be passed to \link{brformat}.}
}
\description{
This is a wrapper around \link{brformat}.
}
