% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rsqrt}
\alias{rsqrt}
\title{Reciprocal of the square root of a number}
\usage{
rsqrt(x)
}
\arguments{
\item{x}{a numeric vector or array}
}
\description{
Reciprocal of the square root of a number
}
\examples{

rsqrt(4)
rsqrt(1:10)
}
