% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Involutivo.R
\name{sensibilidade_vv}
\alias{sensibilidade_vv}
\title{Sensibilidade-Fluxo de Vendas}
\usage{
sensibilidade_vv(range, cc, wc, vgv, wv, bdi_i, bdi_c, cor, tma)
}
\description{
Sensibilidade-Fluxo de Vendas
}
\examples{
wv_otimista <- c(0, 0, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2)/20
wv_pessimista <- c(0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)/20
range_vv <- list(Pessimista = wv_pessimista, Provavel = wv, Otimista = wv_otimista)
s_vv <- sensibilidade_vv(range_vv,
                         cc = cc, wc = wc,
                         vgv = vgv,
                         bdi_i = bdi_i, bdi_c = bdi_c,
                         cor = cor, tma = tma)
s_vv
ggplot(s_vv, aes(x = Situacao, y = vpl)) +
  geom_col() +
  scale_y_continuous(labels = scales::dollar) +
  labs(title = "Sensibilidade à variação da velocidade de vendas")
}
